Feature: Completed step 6 with localstorage
Didn't use the core element localstorage because for some reason I can't get a map item to show in html when it is inside an observable list. When I use an object it appears to work just fine. The core element localstorage doesn't allow me to decode the object an put its items in a Note object, so I opted to do it manually.